requirements.txt
# Web Automation & Image Retrieval
playwright==1.40.0
requests==2.31.0
pillow==10.1.0

# Layout Analysis
layoutparser[layoutmodels]==0.3.4
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
torch==2.1.2
torchvision==0.16.2
opencv-python==4.9.0.80
pdf2image==1.16.3

# OCR Engines
pytesseract==0.3.10
paddleocr==2.7.0.3
paddlepaddle==2.6.0

# Vision & Language Models
openai==1.6.1
google-generativeai==0.3.2
anthropic==0.8.1

# Google Notebook Integration
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0

# Utilities
pydantic==2.5.3
python-dotenv==1.0.0
tqdm==4.66.1
loguru==0.7.2
pyyaml==6.0.1

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
